Move process implementation to package csp.
