+++ pkg/client/client.go
@@ -9,8 +9,8 @@ import (
 	casm "github.com/wetware/casm/pkg"
 	"github.com/wetware/casm/pkg/cluster"
 	"github.com/wetware/casm/pkg/debug"
+	"github.com/wetware/ww/pkg/csp"
 	"github.com/wetware/ww/pkg/host"
-	"github.com/wetware/ww/pkg/process"
 	"github.com/wetware/ww/pkg/pubsub"
 )
 
@@ -54,7 +54,7 @@ func (n *Node) Debug(ctx context.Context) (debug.Debugger, capnp.ReleaseFunc) {
 	return n.Host(ctx).Debug(ctx)
 }
 
-func (n *Node) Executor(ctx context.Context) (process.Executor, capnp.ReleaseFunc) {
+func (n *Node) Executor(ctx context.Context) (csp.Executor, capnp.ReleaseFunc) {
 	return n.Host(ctx).Executor(ctx)
 }
