+++ pkg/csp/executor.go
@@ -1,4 +1,4 @@
-package process
+package csp
 
 import (
 	"context"
