+++ pkg/csp/testdata/main.wasm
